Add OpenRouter AI adapter configuration
