+++ package.json
@@ -62,6 +62,7 @@
     "undici": "^6.19.8"
   },
   "dependencies": {
+    "@ai-sdk/openai": "^3.0.10",
     "@dagrejs/dagre": "^1.1.8",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
@@ -89,6 +90,7 @@
     "@radix-ui/react-tooltip": "^1.2.8",
     "@tanstack/react-query": "^5.90.16",
     "@types/d3-hierarchy": "^3.1.7",
+    "ai": "^6.0.34",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "d3-hierarchy": "^3.1.2",
